version working
user can now search more then 11,000 road objects but only an amount of those are showing on the tableview, for default it's 1,000 road objects showing, but user can changes this limit number if amount of road objects slider widget change.

But road objects amount in the map is loaded full size, that means if user jus want to fecth 1,000 object on tableview and skiltplata in Haugesund is 5,000 object amount, then in tableview will show 1,000 or what user's has had configure to but on map will show 5,000.