Bug Fixed
Bug Fixed when user checking status and expired time is over, meaning that the Plug-In last time loging sessionwas 8 hours ago, and user try to check the endringer /changes status and in an API fetch level gets 404 not found, because of the token expiration.

to fix this now before checking status, Plug-In checks if the current session is expired and if it's then just re-login and add the new id token to all of the sended endringer from before, but only the tokens not endrings id and user must not send again, but re-open the check status windows.