Bug Fixed: onAnySelectedItem and get_related_parent methods
onAnySelectedItem: When user select a road object, then it will populate or not if road object has or not a parent road object id and type.

get_related_parent: get a parsed parent road object and turn some flags on or off, but now if clicked child road object has more then one parent on that fetched, then it will parsed the last one and if not, it will parsed only the first parent.